test: Enhance BankBalanceSection and BankView test coverage
